requirements.txt
streamlit
pandas
plotly
scikit-learn
rank_bm25
pdfplumber
docx2txt
pytest
pytest-asyncio
python-dotenv
scipy
faiss-cpu
strands-agents
strands-agents-tools
openai
pydantic
pydantic-settings
google-genai
rank_bm25
